uvr-package {uvr}	R Documentation

R Interface to the 'uvr' Package Manager

Description

Provides R functions that wrap the uvr command-line tool for fast, reproducible R package management. Use uvr to manage project dependencies with a lockfile, similar to how uv works for Python.

Main functions:

init

Initialize a new uvr project

add

Add packages to the project

remove_pkgs

Remove packages from the project

sync

Install all packages from the lockfile

lock

Re-resolve dependencies and update the lockfile

run

Run an R script in the project environment

install_uvr

Install the uvr CLI binary

update_uvr

Update both R package and CLI binary in one call

update_pkgs

Lock + sync to update installed packages to the latest allowed versions

r_install, r_list, r_use, r_pin

Manage R versions

Typical workflow:

library(uvr)

# Start a new project
init()

# Add packages
add("tidyverse")
add("DESeq2", bioc = TRUE)

# After pulling changes, sync the library
sync()

# Upgrade all packages
lock(upgrade = TRUE)
sync()

See Also

https://github.com/nbafrank/uvr

[Package uvr version 0.1.2 Index]